Bootstrap up to TL2 with Bazel
This allows Bazel to compile the TL2 compiler using TL1 compiler using
TL0 compiler written in C.

Unfortunately, the way TL2 compiler runs is that it creates the generated files
in the same directory as the input (without accepting a parameter) and Bazel
can't handle that nicely without a lot of effort. I think this is still
merge-worthy to have the first few steps automated.